Fix update star account in manager
