+++ LingoClash/LingoClash/Model/Records/StarAccountData.swift
@@ -5,8 +5,6 @@
 //  Created by Ai Ling Hong on 12/4/22.
 //
 
-import Foundation
-
 struct StarAccountData {
     var id: Identifier
     let owner_id: Identifier
@@ -15,13 +13,3 @@ struct StarAccountData {
 
 extension StarAccountData: Record {}
 
-struct StarTransactionData {
-    var id: Identifier
-    let account_id: Identifier
-    let amount: Int
-    let createdAt: Date
-    let debitOrCredit: DebitOrCredit
-    let description: String
-}
-
-extension StarTransactionData: Record {}
